* Modified memory handling in JPEGCompressor.cc to better handle   images where the compressed version may be bigger than the   original (for example at very high quality levels). * Changes to CVT.cc to work with new JPEGCompressor code and removal   of chunked encoding header directive from CVT.cc. * Fixed bug in CVT when specifying both WID and HEI. * Fixed compiler warning in IIPResponse.cc. * Added man page. * Added missing definition of get_nprocs() function to Kakadu.cc   for Mac OSX and FreeBSD.

+++ src/IIPResponse.cc
@@ -102,7 +102,7 @@ void IIPResponse::addResponse( const char* c, int a, int b ){
 void IIPResponse::setError( const string& code, const string& arg ){
 
   char tmp[32];
-  snprintf( tmp, 32, "Error/%ld:%s %s", code.length() + arg.length() + 1, code.c_str(), arg.c_str() );
+  snprintf( tmp, 32, "Error/%ld:%s %s", (long)(code.length() + arg.length() + 1), code.c_str(), arg.c_str() );
   error += tmp + eof;
 }
